further wiring of new services
