+++ app/services/rabbitmq/RabbitMQMessageService.scala
@@ -1,10 +1,10 @@
 package services
 
 import java.io.IOException
-import java.net.URLEncoder
+import java.net.{URI, URLEncoder}
 import java.text.SimpleDateFormat
 
-import akka.actor.{Actor, ActorRef, Props}
+import akka.actor.{Actor, ActorRef, PoisonPill, Props}
 import com.ning.http.client.Realm.AuthScheme
 import com.rabbitmq.client.AMQP.BasicProperties
 import com.rabbitmq.client.{Channel, Connection, ConnectionFactory, DefaultConsumer, Envelope, QueueingConsumer}
@@ -24,12 +24,11 @@ import scala.util.Try
  * Send/get messages from a message bus
  */
 
-trait RabbitMQMessageService extends MessageService {
+class RabbitMQMessageService extends MessageService {
   var channel: Option[Channel] = None
   var connection: Option[Connection] = None
   var factory: Option[ConnectionFactory] = None
 
-  var cancellationQueue: Option[ActorRef] = None
   var event_filter: Option[ActorRef] = None
   var extractorsHeartbeats: Option[ActorRef] = None
 
@@ -51,7 +50,22 @@ trait RabbitMQMessageService extends MessageService {
     if (channel.isDefined) return true
     if (!factory.isDefined) return true
 
+    // Formerly onStart() block
     val configuration = play.api.Play.configuration
+    rabbitmquri = configuration.getString("clowder.rabbitmq.uri").getOrElse("amqp://guest:guest@localhost:5672/%2f")
+    exchange = configuration.getString("clowder.rabbitmq.exchange").getOrElse("clowder")
+    mgmtPort = configuration.getString("clowder.rabbitmq.managmentPort").getOrElse("15672")
+    Logger.debug("uri= "+ rabbitmquri)
+    try {
+      val uri = new URI(rabbitmquri)
+      factory = Some(new ConnectionFactory())
+      factory.get.setUri(uri)
+    } catch {
+      case t: Throwable => {
+        factory = None
+        Logger.error("Invalid URI for RabbitMQ", t)
+      }
+    }
 
     try {
       val protocol = if (factory.get.isSSL) "https://" else "http://"
@@ -152,7 +166,47 @@ trait RabbitMQMessageService extends MessageService {
   }
 
   /** Close connection to broker. **/
-  def close()
+  def close() {
+    Logger.debug("Closing connection")
+    restURL = None
+    vhost = ""
+    username = ""
+    password = ""
+    if (channel.isDefined) {
+      Logger.debug("Channel closing")
+      try {
+        channel.get.close()
+      } catch {
+        case e: Exception => Logger.error("Error closing channel.", e)
+      }
+      channel = None
+    }
+    if (connection.isDefined) {
+      Logger.debug("Connection closing")
+      try {
+        connection.get.close()
+      } catch {
+        case e: Exception => Logger.error("Error closing connection.", e)
+      }
+      connection = None
+    }
+    if (event_filter.isDefined) {
+      event_filter.get ! PoisonPill
+      event_filter = None
+    }
+    if (extractQueue.isDefined) {
+      extractQueue.get ! PoisonPill
+      extractQueue = None
+    }
+    if (cancellationQueue.isDefined) {
+      cancellationQueue.get ! PoisonPill
+      cancellationQueue = None
+    }
+    if (extractorsHeartbeats.isDefined) {
+      extractorsHeartbeats.get ! PoisonPill
+      extractorsHeartbeats = None
+    }
+  }
 
   /** Submit a message to default broker. */
   def submit(message: ExtractorMessage) = {
@@ -178,51 +232,98 @@ trait RabbitMQMessageService extends MessageService {
     }
   }
 
-  def getRestEndPoint(path: String): Future[Response]
+  private def getRestEndPoint(path: String): Future[Response] = {
+    connect
+
+    restURL match {
+      case Some(x) => {
+        val url = x + path
+        Logger.trace("RESTURL: "+ url)
+        WS.url(url).withHeaders("Accept" -> MimeTypes.JSON).withAuth(username, password, AuthScheme.BASIC).get()
+      }
+      case None => {
+        Logger.warn("Could not get bindings")
+        Future.failed(new IOException("Not connected"))
+      }
+    }
+  }
 
   /**
    * Get the exchange list for a given host
    */
-  def getExchanges : Future[Response]
+  def getExchanges : Future[Response] = {
+    connect
+    getRestEndPoint("/api/exchanges/" + vhost )
+  }
 
   /**
    * get list of queues attached to an exchange
    */
-  def getQueuesNamesForAnExchange(exchange: String): Future[Response]
+  def getQueuesNamesForAnExchange(exchange: String): Future[Response] = {
+    connect
+    getRestEndPoint("/api/exchanges/"+ vhost +"/"+ exchange +"/bindings/source")
+  }
 
   /**
    * Get the binding lists (lists of routing keys) from the rabbitmq broker
    */
-  def getBindings: Future[Response]
+  def getBindings: Future[Response] = {
+    getRestEndPoint("/api/bindings")
+  }
 
   /**
    * Get Channel list from rabbitmq broker
    */
-  def getChannelsList: Future[Response]
+  private def getChannelsList: Future[Response] = {
+    getRestEndPoint("/api/channels")
+  }
 
   /**
    * Get queue details for a given queue
    */
-  def getQueueDetails(qname: String): Future[Response]
+  private def getQueueDetails(qname: String): Future[Response] = {
+    connect
+    getRestEndPoint("/api/queues/" + vhost + "/" + qname)
+  }
 
   /**
    * Get queue bindings for a given host and queue from rabbitmq broker
    */
-  def getQueueBindings(qname: String): Future[Response]
+  private def getQueueBindings(qname: String): Future[Response] = {
+    connect
+    getRestEndPoint("/api/queues/" + vhost + "/" + qname + "/bindings")
+  }
 
   /**
    * Get Channel information from rabbitmq broker for given channel id 'cid'
    */
-  def getChannelInfo(cid: String): Future[Response]
+  private def getChannelInfo(cid: String): Future[Response] = {
+    getRestEndPoint("/api/channels/" + cid)
+  }
 
-  def cancelPendingSubmission(id: UUID, queueName: String, msg_id: UUID)
+  def cancelPendingSubmission(id: UUID, queueName: String, msg_id: UUID): Unit = {
+    connect
+    cancellationQueue match {
+      case Some(x) => x ! new CancellationMessage(id, queueName, msg_id)
+      case None => Logger.warn("Could not send message over RabbitMQ")
+    }
+  }
 
   /**
    * a helper function to get user email address from user's request api key.
    * @param requestAPIKey user request apikey
    * @return a list of email address
    */
-  def getEmailNotificationEmailList(requestAPIKey: Option[String]): List[String]
+  private def getEmailNotificationEmailList(requestAPIKey: Option[String]): List[String] = {
+    (for {
+      apiKey <- requestAPIKey
+      user <- userService.findByKey(apiKey)
+      email <- user.email
+    } yield {
+      Logger.debug(s"[getEmailNotificationEmailList] $email")
+      List[String](email)
+    }).getOrElse(List[String]())
+  }
 
   /**
    * loop through the queue and dispatch the message via the routing key.
@@ -231,8 +332,189 @@ trait RabbitMQMessageService extends MessageService {
    * @param channel                    the channel connecting to the rabbitmq
    * @param cancellationSearchTimeout  the timeout of downloading the requests from the rabbitmq
    */
-  def resubmitPendingRequests(cancellationQueueConsumer: QueueingConsumer, channel: Channel, cancellationSearchTimeout: Long)
+  def resubmitPendingRequests(cancellationQueueConsumer: QueueingConsumer, channel: Channel, cancellationSearchTimeout: Long) = {
+    var loop = true
+    while( loop ) {
+      val delivery: QueueingConsumer.Delivery = cancellationQueueConsumer.nextDelivery(cancellationSearchTimeout)
+      delivery match {
+        case null => {
+          Logger.debug(s"[CANCELLATION] read $cancellationDownloadQueueName timeout, exit the looping on the cancellation download queue")
+          loop = false
+        }
+        case _ => {
+          val body = delivery.getBody()
+          val delivery_tag = delivery.getEnvelope.getDeliveryTag
+          val body_text = new String(body)
+          val json = Json.parse(body_text)
+          val routing_key: String = (json \ "routing_key").as[String]
+          val basicProperties = new BasicProperties().builder()
+            .contentType(MimeTypes.JSON)
+            .deliveryMode(2)
+            .build()
+          try {
+            val request_id = (json \ "msgid").asOpt[String] match {
+              case Some(id)=> Some(UUID(id))
+              case None => None
+            }
 
+            channel.basicAck(delivery_tag, false)
+            Logger.debug(s"[CANCELLATION] ACK $request_id to be removed from $cancellationDownloadQueueName")
+            channel.basicPublish(exchange, routing_key, true, basicProperties, body)
+            Logger.debug(s"[CANCELLATION] resubmit to $request_id, $routing_key, $body_text ")
+          } catch {
+            case e: Exception => {
+              Logger.error(s"[CANCELLATION] failed to publish, $routing_key", e)
+            }
+          }
+        }
+      }
+    }
+  }
+
+}
+
+/**
+ * First, it will connect to the target rabbitmq queue to download each pending submission request and search the
+ * cancellation submission by comparing the message id of each pending submission request.
+ * This search will terminate when any of the following condition stands:
+ *   1. the cancellation submission has been found.
+ *   2. within the certain timeout, the target queue has no new pending submission.
+ *   3. the number of downloaded pending submission requests exceeds the Threshold.
+ *
+ * each downloaded submission(except the cancellation submission) will be forwarded to a named rabbitmq queue.
+ *
+ * Second, when the searching is terminated, it will remove and resubmit each submission from the named rabbitmq queue
+ * to the extractor queue(s) based on the routing key of each submission.
+ *
+ * @param exchange                       the exchange of the rabbitmq
+ * @param connection                     the connection to the rabbitmq
+ * @param cancellationDownloadQueueName  the queue name of the cancellation downloaded queue
+ */
+class PendingRequestCancellationActor(exchange: String, connection: Option[Connection], cancellationDownloadQueueName: String,
+                                      cancellationSearchTimeout: Long) extends Actor {
+  val configuration = play.api.Play.configuration
+  val CancellationSearchNumLimits: Integer = configuration.getString("submission.cancellation.search.numlimits").getOrElse("100").toInt
+  def receive = {
+    case CancellationMessage(id, queueName, msg_id) => {
+      val extractions: ExtractionService = DI.injector.getInstance(classOf[ExtractionService])
+      val messages: MessageService = DI.injector.getInstance(classOf[MessageService])
+
+      val dateFormatter = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ssX")
+      var startDate = new java.util.Date()
+      var user_id =  User.anonymous.id
+      val job_id: Option[UUID] = extractions.get(msg_id) match {
+        case Some(extraction) => {
+          user_id = extraction.user_id
+          extraction.job_id
+        }
+        case None => {
+          Logger.warn("Failed to lookup jobId.. no extraction message found with id=" + msg_id)
+          None
+        }
+      }
+      extractions.insert(Extraction(UUID.generate(), id, job_id, queueName, "Cancel Requested", startDate, None, user_id))
+
+      val channel: Channel = connection.get.createChannel()
+      //1. connect to the target rabbitmq queue
+      val queueConsumer: QueueingConsumer = new QueueingConsumer(channel)
+      val queueConsumerTag: String = channel.basicConsume(queueName, false, queueConsumer)
+      val pendingMessages: Integer = channel.queueDeclarePassive(queueName).getMessageCount()
+      val maxSearchNum = math.max(CancellationSearchNumLimits, pendingMessages)
+      Logger.debug(s"[CANCELLATION] receive the cancellation request, $queueName, $msg_id search num limits: $maxSearchNum timeout: $cancellationSearchTimeout")
+      var loop: Boolean = true
+      // 2. parse each pending request and search for the cancellation request
+      var counts = 0
+      var foundCancellationRequest = false
+      while( loop ) {
+        val delivery: QueueingConsumer.Delivery = queueConsumer.nextDelivery(cancellationSearchTimeout)
+        delivery match {
+          case null => {
+            Logger.debug(s"[CANCELLATION] read, $queueName timeout, exit the searching of cancellation submission")
+            loop = false
+          }
+          case _ => {
+            val body = delivery.getBody()
+            val delivery_tag = delivery.getEnvelope.getDeliveryTag
+            val body_text = new String(body)
+            val json = Json.parse(body_text)
+
+            val request_id = (json \ "msgid").asOpt[String] match {
+              case Some(id)=> Some(UUID(id))
+              case None => None
+            }
+
+            if(request_id.isDefined && msg_id.toString() == request_id.get.toString()) {
+              Logger.debug(s"found cancellation request and then skip $request_id")
+              loop = false
+              foundCancellationRequest = true
+            } else {
+              // upload parsed pending requests to the cancellation download queue
+              try {
+                val basicProperties = new BasicProperties().builder()
+                  .contentType(MimeTypes.JSON)
+                  .deliveryMode(2)
+                  .build()
+                channel.basicPublish("", cancellationDownloadQueueName, basicProperties, body)
+                Logger.debug(s"[CANCELLATION] publish $request_id to the queue: $cancellationDownloadQueueName, $body_text")
+              } catch {
+                case e: Exception => {
+                  Logger.error(s"[CANCELLATION] failed to publish to queue: $cancellationDownloadQueueName", e)
+                }
+              }
+            }
+            try {
+              channel.basicAck(delivery_tag, false)
+              Logger.debug(s"[CANCELLATION] ACK $request_id to be removed from $queueName")
+            } catch {
+              case e: Exception => {
+                Logger.error(s"[CANCELLATION] failed to ACK $request_id, $body_text", e)
+              }
+            }
+          }
+        }
+        counts += 1;
+        if (counts > maxSearchNum) {
+          loop = false
+        }
+      }
+      // update extraction event
+      startDate = new java.util.Date()
+      if(foundCancellationRequest) {
+        extractions.insert(Extraction(UUID.generate(), id, job_id, queueName, "Cancel Success", startDate, None, user_id))
+      } else {
+        extractions.insert(Extraction(UUID.generate(), id, job_id, queueName, "Cancel Failed", startDate, None, user_id))
+      }
+
+      try {
+        channel.basicCancel(queueConsumerTag)
+
+      } catch {
+        case e: Exception => {
+          Logger.error(s"[CANCELLATION] failed to cancel, $queueConsumerTag", e)
+        }
+      }
+
+      //3. resubmit pending requests from cancellation download queue to the target queue.
+      val cancellationQueueConsumer: QueueingConsumer = new QueueingConsumer(channel)
+      val cancellationQueueConsumerTag: String = channel.basicConsume(cancellationDownloadQueueName, false, cancellationQueueConsumer)
+
+      messages.resubmitPendingRequests(cancellationQueueConsumer, channel, cancellationSearchTimeout)
+
+      try {
+        channel.basicCancel(cancellationQueueConsumerTag)
+        channel.close()
+      } catch {
+        case e: Exception => {
+          Logger.error(s"[CANCELLATION] failed to cancel $queueConsumerTag", e)
+        }
+      }
+      Logger.debug(s"[CANCELLATION] finish cancellation request $queueName, $msg_id")
+
+    }
+    case _ => {
+      Logger.error("[CANCELLATION] Unknown message type submitted.")
+    }
+  }
 }
 
 /**
@@ -243,6 +525,7 @@ class PublishDirectActor(channel: Channel, replyQueueName: String) extends Actor
   val appHttpPort = play.api.Play.configuration.getString("http.port").getOrElse("")
   val appHttpsPort = play.api.Play.configuration.getString("https.port").getOrElse("")
   val clowderurl = play.api.Play.configuration.getString("clowder.rabbitmq.clowderurl")
+  val messages: MessageService = DI.injector.getInstance(classOf[MessageService])
 
   def receive = {
     case ExtractorMessage(msgid, fileid, jobid, notifies, intermediateId, host, key, metadata, fileSize, datasetId, flags, secretKey, routingKey,
@@ -305,9 +588,7 @@ class PublishDirectActor(channel: Channel, replyQueueName: String) extends Actor
       } catch {
         case e: Exception => {
           Logger.error("Error connecting to rabbitmq broker", e)
-          current.plugin[RabbitmqPlugin].foreach {
-            _.close()
-          }
+          messages.close
         }
       }
     }
