+++ app/views/file.scala.html
@@ -402,8 +402,7 @@
 							objectType="file"><span class="glyphicon glyphicon-star-empty"></span> Unfollow
 							</button>
 						}
-                        @if(Permission.checkPermission(Permission.EditFile, ResourceRef(ResourceRef.file, file.id))
-                                && play.api.Play.current.plugin[services.RabbitmqPlugin].isDefined) {
+                        @if(Permission.checkPermission(Permission.EditFile, ResourceRef(ResourceRef.file, file.id))) {
                             <a href="@routes.Extractors.submitFileExtraction(UUID(id))" class="btn btn-link" >
                                 <span class="glyphicon glyphicon-send"></span>
                                 Submit for extraction
