+++ conf/play.plugins
@@ -1,6 +1,5 @@
 9990:services.mongodb.MongoSalatPlugin
 9991:com.typesafe.plugin.CommonsMailerPlugin
-#9992:services.RabbitmqPlugin
 9994:services.mongodb.MongoDBAuthenticatorStore
 9995:securesocial.core.DefaultIdGenerator
 9996:securesocial.core.providers.utils.DefaultPasswordValidator
